refactor: add RemoteAttendee
