+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/model/Event.kt
@@ -16,41 +16,19 @@ data class RemoteEvent(
     val remindAt: Long,
     val host: String,
     val isUserEventCreator: Boolean,
-    val attendees: List<Attendee>,
+    val attendees: List<RemoteAttendee>,
     val photos: List<Photo>,
 )
 
 @Serializable
-data class Attendee(
+data class RemoteAttendee(
     val email: String,
     val fullName: String,
     val userId: String,
     val eventId: String,
     val isGoing: Boolean,
     val remindAt: Long,
-) {
-    companion object {
-        val DUMMY_LIST =
-            listOf(
-                Attendee(
-                    email = "attendee1@example.com",
-                    fullName = "Attendee 1 Name",
-                    userId = "user123",
-                    eventId = "event123",
-                    isGoing = true,
-                    remindAt = 1678886400000,
-                ),
-                Attendee(
-                    email = "attendee2@example.com",
-                    fullName = "Attendee 2 Name",
-                    userId = "user456",
-                    eventId = "event123",
-                    isGoing = false,
-                    remindAt = 1678886400000,
-                ),
-            )
-    }
-}
+)
 
 @Serializable
 data class Photo(
